+++ backend/app.py
@@ -9,6 +9,7 @@ from loguru import logger
 
 from backend.api.health import router as health_router
 from backend.api.products import router as products_router
+from backend.api.recommendations import router as recommendations_router
 from backend.api.stores import stores_router, users_router
 from backend.api.watches import router as watches_router
 from backend.config import SERVICE_NAME, backend_settings
@@ -42,3 +43,4 @@ app.include_router(products_router, prefix="/api")
 app.include_router(stores_router, prefix="/api")
 app.include_router(users_router, prefix="/api")
 app.include_router(watches_router, prefix="/api")
+app.include_router(recommendations_router, prefix="/api")